+++ Dictionary/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+- Value replacements in ObservableDictionaryBase subclasses no longer
+  trigger a PropertyChangedEvent on Keys.
+### Fixed
+- Clearing ObservableDictionaryBase subclasses no longer throws an
+  exception due to bad NotifyCollectionChangedEventArgs parameters.
+
 ## 0.2.0
 ### Added
 - This change log